TVB-3132: resolve connectivity zip import error

+++ tvb_framework/tvb/core/neotraits/db.py
@@ -94,7 +94,7 @@ class HasTraitsIndex(Base):
 
 
 def ensure_float(data):
-    return numpy.float64(data)
+    return float(data)
 
 
 def ensure_int(data):
